test(nav): migrate tests
